cleaned up validatorama

+++ src/client/scripts/esm/util/validatorama.js
@@ -11,28 +11,41 @@
 
 import docutil from "./docutil.js";
 
-/** Our username, if we are logged in. @type {string | undefined} */
-let username;
 
-/** Access token for authentication, if we are logged in. @type {string | undefined} */
-let accessToken;
-
-/** Last refresh time of the access token, in milliseconds. @type {number | undefined} */
-let lastRefreshTime;
-
-/** Expiration time for the token in milliseconds. @type {number} */
+const minTimeToRenewSession = 1000 * 60 * 60 * 24; // 1 day
+/** Expiration time for the access tokens. @type {number} */
 const TOKEN_EXPIRE_TIME_MILLIS = 1000 * 60 * 15; // 15 minutes
 
+
+let memberInfo = {
+	signedIn: false,
+	user_id: undefined,
+	username: undefined,
+	issued: undefined,
+	expires: undefined,
+}
+
+let tokenInfo = {
+	/** Access token for authentication, if we are logged in AND have requested one! @type {string | undefined} */
+	accessToken: undefined,
+	/** Last refresh time of the access token, in milliseconds. @type {number | undefined} */
+	lastRefreshTime: undefined,
+}
 /** Cushion time in milliseconds before considering the token expired. @type {number} */
 const CUSHION_MILLIS = 10_000;
 
-let reqIsOut = false;
+(function init() {
+	document.addEventListener('logout', event => { // Custom-event listener. Often fired when a web socket connection closes due to us logging out.
+		memberInfo = { signedIn: false };
+		tokenInfo = {};
+	});
+	
+	// Sets our memberInfo properties if we are logged in
+	readMemberInfoCookie();
+	// Most of the time we don't need an immediate access token
+	// refreshToken();
+})();
 
-document.addEventListener('logout', event => { // Custom-event listener. Often fired when a web socket connection closes due to us logging out.
-	username = undefined;
-	accessToken = undefined;
-	lastRefreshTime = undefined;
-});
 
 /**
  * Checks if the access token is expired or near-expiring.
@@ -44,17 +57,17 @@ document.addEventListener('logout', event => { // Custom-event listener. Often f
 async function getAccessToken() {
 	if (reqIsOut) await waitUntilInitialRequestBack();
 
-	if (!areWeLoggedIn()) return;
+	if (!memberInfo.signedIn) return;
 
 	const currentTime = Date.now();
-	const timeSinceLastRefresh = currentTime - (lastRefreshTime || 0);
+	const timeSinceLastRefresh = currentTime - (tokenInfo.lastRefreshTime || 0);
 
 	// Check if token is expired or near expiring
-	if (!accessToken || timeSinceLastRefresh > (TOKEN_EXPIRE_TIME_MILLIS - CUSHION_MILLIS)) {
+	if (!tokenInfo.accessToken || timeSinceLastRefresh > (TOKEN_EXPIRE_TIME_MILLIS - CUSHION_MILLIS)) {
 		await refreshToken();
 	}
 
-	return accessToken;
+	return tokenInfo.accessToken;
 }
 
 /**
@@ -66,39 +79,39 @@ async function getAccessToken() {
  * @returns {Promise<void>} Resolves when the token refresh process is complete.
  */
 async function refreshToken() {
-	reqIsOut = true;
-	let OK = false;
-
 	try {
 		const response = await fetch('/api/get-access-token', {
 			method: 'POST', // Ensure it's a POST request
 			headers: {
-				'Content-Type': 'application/json', // Add the appropriate headers if needed
+				'Content-Type': 'application/json',
 			},
 		});
-		OK = response.ok;
 
 		const result = await response.json();
 
-		if (OK) { // Refresh token (from cookie) accepted!
-			accessToken = docutil.getCookieValue('token'); // Read access token from cookie
-			if (!accessToken) console.error("Token not found in the cookie!");
-			lastRefreshTime = Date.now(); // Update the last refresh time
-		} else { // 403 of 500 error
+		if (response.ok) { // Refresh token (from cookie) accepted!
+			const accessToken = docutil.getCookieValue('token'); // Read access token from cookie
+			if (!accessToken) throw new Error('Token cookie not found!');
+			tokenInfo = { accessToken, lastRefreshTime: Date.now() }
+
+			// Delete the token cookie after reading it
+			docutil.deleteCookie('token');
+			
+			// It's possible the server renewed our session. Let's read the memberInfo cookie again!
+			readMemberInfoCookie();
+
+			// Dispatch event to inform other parts of the app that we are logged in.
+			// document.dispatchEvent(new CustomEvent('login'));
+
+		} else { // 403 or 500 error   Likely not signed in!
 			console.log(`Server: ${result.message}`);
 			docutil.deleteCookie('memberInfo');
+			memberInfo = { signedIn: false };
 		}
-
-		// Delete the token cookie after reading it
-		docutil.deleteCookie('token');
-		refreshOurUsername();
 
 	} catch (error) {
 		console.error('Error occurred during token refresh:', error);
-	} finally {
-		reqIsOut = false;
-		// Dispatch event to inform other parts of the app that validation is complete
-		document.dispatchEvent(new CustomEvent('validated'));
+		readMemberInfoCookie();
 	}
 }
 
@@ -107,29 +120,19 @@ async function refreshToken() {
  * if we have a refreshed token cookie, to grab our
  * username and user_id properties if we are signed in.
  */
-function refreshOurUsername() {
+function readMemberInfoCookie() {
 	// Read the member info from the cookie
 	// Get the URL-encoded cookie value
 	// JSON objects can't be string into cookies because cookies can't hold special characters
 	const encodedMemberInfo = docutil.getCookieValue('memberInfo'); 
 	if (!encodedMemberInfo) {
-		username = undefined;
+		memberInfo = { signedIn: false }
 		return; // No cookie, not signed in.
 	}
 	// Decode the URL-encoded string
 	const memberInfoStringified = decodeURIComponent(encodedMemberInfo);
-	const memberInfo = JSON.parse(memberInfoStringified); // { user_id, username, issued, expires }
-
-	username = memberInfo.username;
-}
-
-/**
- * Waits until the initial request for an access token is completed.
- */
-async function waitUntilInitialRequestBack() {
-	while (reqIsOut) {
-		await new Promise(resolve => setTimeout(resolve, 100));
-	}
+	memberInfo = JSON.parse(memberInfoStringified); // { user_id, username, issued, expires }
+	memberInfo.signedIn = true;
 }
 
 /**
@@ -137,7 +140,7 @@ async function waitUntilInitialRequestBack() {
  * @returns {boolean} True if logged in, false otherwise.
  */
 function areWeLoggedIn() {
-	return username !== undefined;
+	return memberInfo.signedIn;
 }
 
 /**
@@ -145,15 +148,13 @@ function areWeLoggedIn() {
  * @returns {string | undefined} The username, or undefined if not logged in.
  */
 function getOurUsername() {
-	return username;
+	return memberInfo.signedIn ? memberInfo.username : undefined;
 }
 
-refreshOurUsername();
-refreshToken();
+
 
 // Export these methods to be used by other scripts
 export default {
-	refreshToken,
 	waitUntilInitialRequestBack,
 	areWeLoggedIn,
 	getOurUsername,
